+++ resources/currentUser.txt
@@ -1,1 +1,1 @@
-kokoka
+sasasa
